Correct changing personal data (Controller/Service/DB)

+++ src/main/java/ru/geekbrains/clientservice/entity/Client.java
@@ -21,7 +21,7 @@ public class Client {
     @Id
     @Column(name = "client_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int clientId;
+    private long clientId;
 
     @Column(name = "client_name")
     private String clientName;
